Correcton to specs
